backend/requirements.txt
# 药品管理系统依赖

# Web框架
Flask==3.0.0
Flask-CORS==4.0.0

# 数据库
SQLAlchemy==2.0.23

# 环境变量
python-dotenv==1.0.0

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# HTTP请求
requests==2.31.0

# 腾讯云SDK（可选，用于腾讯云OCR）
# tencentcloud-sdk-python==3.0.1000

# WSGI服务器
gunicorn==21.2.0
